requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas>=2.1.0
numpy>=1.26.0
pyyaml==6.0.2
httpx==0.27.0
apscheduler==3.10.4
python-multipart==0.0.9
